Bug fix: Do not exclude tests with inherited RunWith annotations
sbt does not match an AnnotationFingerprint if the annotation is on
a superclass, so we must not check for those annotations either
when excluding RunWith-based test classes in JUnitFingerprint.

This also means that an inherited @RunWith alone (which would be enough
for JUnit to run a class properly) is not sufficient for getting the
class discovered by sbt.